Adding WorldGuardApi class
